Crate auraescript

An interpreted infrastructure language built for enterprise platform teams.

AuraeScript is an opinionated and Turing complete client language for an Aurae server. AuraeScript is an alternative to templated YAML for teams to express their applications.

The AuraeScript definition lives in this crate library (lib.rs).

Modules

builtin

The builtin functionality for AuraeScript.

meta

observe

runtime

schedule

Functions

register_stdlib

AuraeScript Standard Library.