Module auraescript::runtime

Modules

core_client

Generated client implementations.

Structs

Cell

A set of containers running on the same kernel as the auraed responding to the request.

CellStatus

Container

Container represents is an OCI compliant container image which can be executed.

ContainerStatus

ContainerStatus is the status of a container after it has been executed.

Executable

Executable is the lowest level of compute that Aurae can execute. A basic process.

ExecutableStatus

ExecutableStatus is only returned after a process completes. Because runtime is a synchronous subsystem this will only return upon a terminated process.

Instance

Instance is a recursive graph structure which holds the meta data for the nested Aurae instances running on a machine.

Pod

Pod is a group of containers running in a spawned Aurae instance.

PodStatus

PodStatus is the status of a completed pod and its subsequent containers.

Runtime

SpawnRequest

Parameters for spawning a new nested Aurae instance.

SpawnResponse

Response of a spawn.

VirtualMachine

A long lived virtual machine which will persist on termination.

VirtualMachineStatus

Status of a terminated virtual machine.

Functions

cmd

exec