Module auraescript::runtime

Modules

cell_service_client

Generated client implementations.

instances_client

Generated client implementations.

pods_client

Generated client implementations.

spawn_client

Generated client implementations.

Structs

AllocateCellRequest

AllocateCellResponse

Cell

/ An isolation resource used to divide a system into smaller resource boundaries.

CellService

Executable

/ The most primitive workload in Aurae, a standard executable process.

ExecutableReference

/ A reference to an executable and it’s parent cell.

FreeCellRequest

FreeCellResponse

StartCellRequest

StartCellResponse

StopCellRequest

StopCellResponse